Move long assets to separate files
