+++ resources/tentative/editors/assets/README.md
@@ -1,0 +1,2 @@
+longscript.js: the script comes from Apple
+longtext.html: This is the first chapter from Moby Dick (instead of the whole book at https://github.com/ueberdosis/tiptap/tree/main/demos/src/Examples/Book)
